requirements.txt
google-generativeai
pdfplumber
pandas
openpyxl
streamlit
python-dotenv
